Start working on make the rune crafting table has a resource cost.
Basic fields created
Function calls made

Next is to add the recipes into Unity.
Then updated the resource cost list to grab the item,s for those recipes.

+++ Assets/Scripts/FireStatusManager.cs
@@ -51,7 +51,7 @@ public class FireStatusManager : MonoBehaviour
         Destroy(statusAffect);
 
         statusAffect = gameObject.AddComponent<ReflectBurning>();
-        statusAffect.CalculateStats(GetComponent<LivingEntities>().GetSkillLevel((int)SkillType.Pyromancy));
+        statusAffect.CalculateStats(GetComponent<LivingEntities>().GetSkillLevel(SkillType.Pyromancy));
     }
 
     public void SetIsBurning(bool state)
